# EnterpriseWarehouse.Wasm/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EnterpriseWarehouse.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Supply": {
      "get": {
        "tags": [
          "Supply"
        ],
        "summary": "Возвращает список всех поставок.",
        "responses": {
          "200": {
            "description": "Список поставок успешно возвращён.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplyDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplyDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SupplyDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Supply"
        ],
        "summary": "Добавляет новую поставку.",
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.SupplyCreateDTO, содержащий информацию о новой поставке.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              }
            }
          },
          "201": {
            "description": "Поставка успешно добавлена."
          },
          "400": {
            "description": "Ошибка при добавлении поставки."
          }
        }
      }
    },
    "/api/Supply/{id}": {
      "get": {
        "tags": [
          "Supply"
        ],
        "summary": "Получает информацию о поставке по её идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор поставки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Поставка найдена и информация успешно возвращена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              }
            }
          },
          "404": {
            "description": "Поставка с указанным идентификатором не найдена."
          }
        }
      },
      "put": {
        "tags": [
          "Supply"
        ],
        "summary": "Обновляет информацию о существующей поставке.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор поставки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект EnterpriseWarehouse.API.DTO.SupplyCreateDTO, содержащий обновлённые данные поставки.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplyCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Поставка успешно обновлена.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplyDTO"
                }
              }
            }
          },
          "404": {
            "description": "Поставка с указанным идентификатором не найдена."
          }
        }
      },
      "delete": {
        "tags": [
          "Supply"
        ],
        "summary": "Удаляет поставку по её идентификатору.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор поставки.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Поставка успешно удалена."
          },
          "404": {
            "description": "Поставка с указанным идентификатором не найдена."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "OrganizationDTO": {
        "required": [
          "address",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO организации"
      },
...
      "ProductDTO": {
        "required": [
          "code",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "code": {
            "type": "integer",
            "description": "Код товара",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Название",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для товара"
      },
...
      "SupplyCreateDTO": {
        "required": [
          "organizationId",
          "productId",
          "quantity",
          "supplyDate"
        ],
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "Идентификатор товара",
            "format": "int32"
          },
          "organizationId": {
            "type": "integer",
            "description": "Идентификатор организации",
            "format": "int32"
          },
          "supplyDate": {
            "type": "string",
            "description": "Дата поставки",
            "format": "date-time"
          },
          "quantity": {
            "type": "integer",
            "description": "Количество поставленного товара",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания поставки"
      },
      "SupplyDTO": {
        "required": [
          "id",
          "organization",
          "product",
          "quantity",
          "supplyDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "product": {
            "$ref": "#/components/schemas/ProductDTO"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationDTO"
          },
          "supplyDate": {
            "type": "string",
            "description": "Дата поставки",
            "format": "date-time"
          },
          "quantity": {
            "type": "integer",
            "description": "Количество поставленного товара",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO поставки"
      }
    }
  }
}